[#3636] IndicatorNavMenu trigger id selection on activeKey current (as panels are no longer forceRendered)
