+++ akvo/rsr/spa/app/modules/editor/section5/styles.scss
@@ -121,6 +121,9 @@
       height: 81px;
       display: flex;
       align-items: center;
+      .ant-collapse-extra{
+        margin-left: auto;
+      }
       .delete-btn-holder{
         overflow: hidden;
         .ant-btn{
